requirements.txt
# Install required libraries with the following command:
#     pip install -r requirements.txt

discord.py ~= 2.3
